fix: add more utils unit tests
