+++ tests/util/utils.ts
@@ -19,12 +19,4 @@ Deno.test({
 
 // delay
 
-Deno.test({
-  name: "[utils] delay 2000 ms",
-  async fn() {
-    const before = Date.now();
-    await delay(2000);
-    const after = Date.now();
-    assertEquals(after - before >= 2000, true);
-  },
-});
+
