+++ testss/delay.test.ts
@@ -1,0 +1,12 @@
+import { delay } from "../mod.ts";
+import { assertEquals } from "./deps.ts";
+
+Deno.test({
+  name: "[utils] delay 2000 ms",
+  async fn() {
+    const before = Date.now();
+    await delay(2000);
+    const after = Date.now();
+    assertEquals(after - before >= 2000, true);
+  },
+});
